backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.1
python-dotenv==1.0.0
joblib==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
httpx==0.24.1
lxml==4.9.3
python-dateutil==2.8.2
selenium==4.14.0
webdriver-manager==3.8.6
